// app/api/src/types/hubspot-true-dialog-association-info.ts
export interface HubspotTrueDialogAccountInfo {
  id: string;
  userEmail: string;
  portalId: string;
  clientHubspotTrueDialogAccountInfo: {
    refreshToken: string;
    trueDialogAccountId: string;
    trueDialogApiKey: string;
    trueDialogApiSecret: string;
  };
  createdAt: string;
  _rid: string;
  _self: string;
  _etag: string;
  _attachments: string;
  _ts: number;
}